update_gpu_drivers: Keep temporary dirs in dry run mode
The flag description says the temporary dirs will be kept in dry run.

BUG=None
TEST=update_gpu_drivers.sh --package mali-drivers-valhall -n
     Then check /tmp/

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3912516
Commit-Queue: Fei Shao <fshao@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>

+++ contrib/update_gpu_drivers.sh
@@ -311,7 +311,9 @@ build_board() {
 cleanup() {
   trap - INT TERM ERR EXIT
 
-  rm -rf "${TEMP_DIR}"
+  if [[ ${FLAGS_dryrun} -eq ${FLAGS_FALSE} ]]; then
+    rm -rf "${TEMP_DIR}"
+  fi
 }
 
 main() {
